refactor: use bitcoin wallet standard feature exports
