+++ packages/connect-bitcoin/package.json
@@ -58,7 +58,7 @@
     "test:watch": "vitest watch"
   },
   "dependencies": {
-    "@metamask/bitcoin-wallet-standard": "^0.2.0",
+    "@metamask/bitcoin-wallet-standard": "^1.2.0",
     "@metamask/connect-multichain": "workspace:^",
     "@wallet-standard/base": "^1.1.0"
   },
