--- playground/browser-playground/src/helpers/bitcoinFeatures/disconnect.ts
@@ -1,13 +1,0 @@
-/** Name of the feature. */
-export const BitcoinDisconnect = 'bitcoin:disconnect';
-
-export type BitcoinDisconnectFeature = {
-  readonly [BitcoinDisconnect]: {
-    readonly version: BitcoinDisconnectVersion;
-    readonly disconnect: BitcoinDisconnectMethod;
-  };
-};
-
-export type BitcoinDisconnectVersion = '1.0.0';
-
-export type BitcoinDisconnectMethod = () => Promise<void>;
