--- playground/browser-playground/src/helpers/bitcoinFeatures/satsConnect.ts
@@ -1,9 +1,0 @@
-import type { BitcoinProvider } from 'sats-connect';
-
-export const BitcoinSatsConnect = 'sats-connect:';
-
-export type BitcoinSatsConnectFeature = {
-  [BitcoinSatsConnect]: {
-    provider: BitcoinProvider;
-  };
-};
